completed CSS for single product page
