+++ client/components/index.js
@@ -9,4 +9,4 @@ export {default as WelcomePage} from './welcome-page'
 export {default as AllProducts} from './products/all-products'
 export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
-export {SingleProduct} from './single-product'
+export {SingleProduct} from './products/single-product'
