--- client/components/single-product.css
@@ -1,10 +1,0 @@
-.add-btn {
-  padding: 20px 30px;
-  color: #ffffff;
-  font-weight: bold;
-  font: 12px graphik--medium, 'Helvetica Neue';
-  letter-spacing: 0.17em;
-  background-color: #4c4c4b;
-  border: 1px solid white;
-  border-radius: 2px;
-}
