+++ client/routes.js
@@ -4,7 +4,7 @@ import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
 import {Login, Signup, UserHome} from './components'
 import {me} from './store'
-import {SingleProduct} from './components/single-product'
+import {SingleProduct} from './components/products/single-product'
 
 /**
  * COMPONENT
